# docs/swagger.yaml
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Seat:
    properties:
      coord_x:
        type: number
      coord_y:
        type: number
      created_at:
        type: string
      id:
        type: string
      row_label:
        type: string
      seat_number:
        type: string
      venue_section_id:
        type: string
    type: object
...
  models.Venue:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      province:
        type: string
      total_capacity:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.VenueSection:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      id:
        type: string
      is_seated:
        type: boolean
      name:
        type: string
      venue_id:
        type: string
    type: object
...
  venue.CreateSeatRequest:
    properties:
      coord_x:
        type: number
      coord_y:
        type: number
      row_label:
        maxLength: 10
        type: string
      seat_number:
        maxLength: 10
        type: string
    required:
    - row_label
    - seat_number
    type: object
  venue.CreateVenueRequest:
    properties:
      address:
        type: string
      city:
        maxLength: 100
        type: string
      country:
        type: string
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      name:
        maxLength: 200
        type: string
      province:
        type: string
      total_capacity:
        type: integer
      type:
        enum:
        - indoor
        - outdoor
        - hybrid
        type: string
    required:
    - address
    - city
    - name
    - total_capacity
    - type
    type: object
  venue.CreateVenueSectionRequest:
    properties:
      capacity:
        type: integer
      is_seated:
        type: boolean
      name:
        maxLength: 100
        type: string
    required:
    - capacity
    - name
    type: object
  venue.UpdateSeatRequest:
    properties:
      coord_x:
        type: number
      coord_y:
        type: number
      row_label:
        maxLength: 10
        type: string
      seat_number:
        maxLength: 10
        type: string
    type: object
  venue.UpdateVenueRequest:
    properties:
      address:
        type: string
      city:
        maxLength: 100
        type: string
      country:
        type: string
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      name:
        maxLength: 200
        type: string
      province:
        type: string
      total_capacity:
        type: integer
      type:
        enum:
        - indoor
        - outdoor
        - hybrid
        type: string
    type: object
  venue.UpdateVenueSectionRequest:
    properties:
      capacity:
        type: integer
      is_seated:
        type: boolean
      name:
        maxLength: 100
        type: string
    type: object
info:
  contact: {}
  description: API for the Konsera ticketing platform
  title: Konsera API
  version: "1.0"
paths:
...
  /venues:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Venue'
            type: array
      security:
      - BearerAuth: []
      summary: List venues
      tags:
      - Venues
    post:
      consumes:
      - application/json
      parameters:
      - description: Venue data
        in: body
        name: venue
        required: true
        schema:
          $ref: '#/definitions/venue.CreateVenueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Venue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create venue
      tags:
      - Venues
  /venues/{venue_id}:
    delete:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete venue
      tags:
      - Venues
    get:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Venue'
      security:
      - BearerAuth: []
      summary: Get venue
      tags:
      - Venues
    put:
      consumes:
      - application/json
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Venue changes
        in: body
        name: venue
        required: true
        schema:
          $ref: '#/definitions/venue.UpdateVenueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Venue'
      security:
      - BearerAuth: []
      summary: Update venue
      tags:
      - Venues
  /venues/{venue_id}/sections:
    get:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.VenueSection'
            type: array
      security:
      - BearerAuth: []
      summary: List venue sections
      tags:
      - Venue Sections
    post:
      consumes:
      - application/json
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section data
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/venue.CreateVenueSectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.VenueSection'
      security:
      - BearerAuth: []
      summary: Create venue section
      tags:
      - Venue Sections
  /venues/{venue_id}/sections/{section_id}:
    delete:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete venue section
      tags:
      - Venue Sections
    get:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VenueSection'
      security:
      - BearerAuth: []
      summary: Get venue section
      tags:
      - Venue Sections
    put:
      consumes:
      - application/json
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      - description: Section changes
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/venue.UpdateVenueSectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VenueSection'
      security:
      - BearerAuth: []
      summary: Update venue section
      tags:
      - Venue Sections
  /venues/{venue_id}/sections/{section_id}/seats:
    get:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Seat'
            type: array
      security:
      - BearerAuth: []
      summary: List seats
      tags:
      - Seats
    post:
      consumes:
      - application/json
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      - description: Seat data
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/venue.CreateSeatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Seat'
      security:
      - BearerAuth: []
      summary: Create seat
      tags:
      - Seats
  /venues/{venue_id}/sections/{section_id}/seats/{seat_id}:
    delete:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seat_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete seat
      tags:
      - Seats
    get:
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seat_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seat'
      security:
      - BearerAuth: []
      summary: Get seat
      tags:
      - Seats
    put:
      consumes:
      - application/json
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: string
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seat_id
        required: true
        type: string
      - description: Seat changes
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/venue.UpdateSeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seat'
      security:
      - BearerAuth: []
      summary: Update seat
      tags:
      - Seats
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"